added a hover for bottom box to show picture when hovered over
again experimenting with JJ's project, does not fork for me but I will
meet with her and see HER final project, it's probably real cool (and
has the appropriate details. This was fun to learn from. And hopefully
I will learn better as I go further and explore and receive
explanations. Amen.